# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite Chauffeur Service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, in_progress, completed, cancelled]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationBody"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Delete/cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reservations/calendar:
    get:
      operationId: getCalendarReservations
      tags: [reservations]
      summary: Get reservations formatted for calendar display
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    Car:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        imageUrl:
          type: string
          nullable: true
        capacity:
          type: integer
        pricePerKm:
          type: number
        baseFee:
          type: number
        driverFee:
          type: number
          description: Extra fee added when chauffeur service is selected
        available:
          type: boolean
        category:
          type: string
          enum: [with_driver, without_driver, both]
        description:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - brand
        - model
        - year
        - capacity
        - pricePerKm
        - baseFee
        - driverFee
        - available
        - category
        - features
        - createdAt
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        available:
          type: boolean
        rating:
          type: number
        languages:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - available
        - rating
        - languages
        - yearsExperience
        - createdAt
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        car:
          $ref: "#/components/schemas/Car"
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        pickupDate:
          type: string
          format: date
        pickupTime:
          type: string
        distanceKm:
          type: number
          nullable: true
        driverFee:
          type: number
          nullable: true
          description: Chauffeur service fee applied to this booking
        totalPrice:
          type: number
          nullable: true
        withDriver:
          type: boolean
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - pickupLocation
        - dropoffLocation
        - pickupDate
        - pickupTime
        - withDriver
        - status
        - createdAt
    CreateReservationBody:
      type: object
      properties:
        carId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        pickupDate:
          type: string
          format: date
        pickupTime:
          type: string
        distanceKm:
          type: number
          nullable: true
        driverFee:
          type: number
          nullable: true
          description: Chauffeur service fee applied to this booking
        totalPrice:
          type: number
          nullable: true
        withDriver:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - customerName
        - customerEmail
        - pickupLocation
        - dropoffLocation
        - pickupDate
        - pickupTime
        - withDriver
    UpdateReservationBody:
      type: object
      properties:
        carId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        pickupDate:
          type: string
          format: date
        pickupTime:
          type: string
        distanceKm:
          type: number
          nullable: true
        driverFee:
          type: number
          nullable: true
        totalPrice:
          type: number
          nullable: true
        withDriver:
          type: boolean
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        notes:
          type: string
          nullable: true
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
        customerName:
          type: string
        carName:
          type: string
          nullable: true
        driverName:
          type: string
          nullable: true
      required:
        - id
        - title
        - date
        - time
        - status
        - customerName
...